fix: refactoring transaction dto generation package
