+++ src/main/java/it/pagopa/ecommerce/commons/documents/TransactionAddReceiptData.java
@@ -1,6 +1,6 @@
 package it.pagopa.ecommerce.commons.documents;
 
-import it.pagopa.generated.transactions.server.model.TransactionStatusDto;
+import it.pagopa.ecommerce.commons.generated.server.model.TransactionStatusDto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import org.springframework.data.mongodb.core.mapping.Document;
