+++ src/main/java/it/pagopa/ecommerce/commons/documents/TransactionAuthorizationStatusUpdateData.java
@@ -1,7 +1,7 @@
 package it.pagopa.ecommerce.commons.documents;
 
-import it.pagopa.generated.transactions.server.model.AuthorizationResultDto;
-import it.pagopa.generated.transactions.server.model.TransactionStatusDto;
+import it.pagopa.ecommerce.commons.generated.server.model.AuthorizationResultDto;
+import it.pagopa.ecommerce.commons.generated.server.model.TransactionStatusDto;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import org.springframework.data.mongodb.core.mapping.Document;
