+++ src/main/java/it/pagopa/ecommerce/commons/utils/TransactionUtils.java
@@ -1,6 +1,6 @@
 package it.pagopa.ecommerce.commons.utils;
 
-import it.pagopa.generated.transactions.server.model.TransactionStatusDto;
+import it.pagopa.ecommerce.commons.generated.server.model.TransactionStatusDto;
 import org.springframework.stereotype.Component;
 
 /**
