+++ src/test/java/it/pagopa/ecommerce/commons/documents/TransactionDocumentTest.java
@@ -3,7 +3,7 @@ package it.pagopa.ecommerce.commons.documents;
 import it.pagopa.ecommerce.commons.TransactionUtils;
 import it.pagopa.ecommerce.commons.domain.TransactionActivated;
 import it.pagopa.ecommerce.commons.domain.TransactionActivationRequested;
-import it.pagopa.generated.transactions.server.model.TransactionStatusDto;
+import it.pagopa.ecommerce.commons.generated.server.model.TransactionStatusDto;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.mockito.junit.jupiter.MockitoExtension;
